+++ src/Bible.Alarm/Common/Interfaces/Media/IScheduleDisplayService.cs
@@ -4,7 +4,7 @@ namespace Bible.Alarm.Common.Interfaces.Media;
 
 public interface IScheduleDisplayService
 {
-    Task<string> GetChapterDisplayNameAsync(long scheduleId, bool force = false);
-    Task<string> GetChapterDisplayNameForBibleReadingAsync(long scheduleId, BibleReadingSchedule bibleReadingSchedule, bool force = false);
+    Task<string> GetChapterDisplayNameAsync(int scheduleId, bool force = false);
+    Task<string> GetChapterDisplayNameForBibleReadingAsync(int scheduleId, BibleReadingSchedule bibleReadingSchedule, bool force = false);
 }
