+++ src/Bible.Alarm/Common/Interfaces/Scheduler/ISchedulePersistenceService.cs
@@ -5,7 +5,7 @@ namespace Bible.Alarm.Common.Interfaces.Scheduler;
 public interface ISchedulePersistenceService
 {
     Task<bool> SaveScheduleAsync(AlarmSchedule schedule, bool isNewSchedule, bool musicUpdated = true, bool bibleReadingUpdated = true);
-    Task DeleteScheduleAsync(long scheduleId);
+    Task DeleteScheduleAsync(int scheduleId);
     Task<AlarmSchedule> GetSampleScheduleAsync();
 }
